Enhance BaseRepository with docs, null checks, new methods
Enhanced `BaseRepository` with additional documentation comments for better readability and maintainability. Added null checks in the constructor for `cache`, `executor`, and `websiteChannelContext` parameters, and set `CacheMinutes` to default to 10 if `options` is null. Introduced a new method `ExecuteContentQuery` for executing content queries, and updated it to call `Executor.GetMappedResult<T>` in preview mode. Added a new property `CachePrefix` with a default value of "base|data". Updated `PageTypeRepository` to use the generic method `.ForContentType<TEntity>` in `ContentItemQueryBuilder`.